+++ performance_test/src/experiment_configuration/experiment_configuration.cpp
@@ -194,6 +194,11 @@ void ExperimentConfiguration::setup(int argc, char ** argv)
     std::vector<std::string> allowedExecStrats;
     allowedExecStrats.push_back("INTER_THREAD");
     allowedExecStrats.push_back("INTRA_THREAD");
+#ifdef PERFORMANCE_TEST_APEX_OS_POLLING_SUBSCRIPTION_ENABLED
+    allowedExecStrats.push_back("APEX_SINGLE_EXECUTOR");
+    allowedExecStrats.push_back("APEX_EXECUTOR_PER_COMMUNICATOR");
+    allowedExecStrats.push_back("APEX_CHAIN");
+#endif
     TCLAP::ValuesConstraint<std::string> allowedExecStratVals(allowedExecStrats);
     TCLAP::ValueArg<std::string> executionStrategyArg("e", "execution-strategy",
       "The execution strategy to use. "